Add files via upload
Program currently crawls all files in the working directory. It uses regular expression matching to make sure it is an appropriate file. Then it combines the first initial of the first name and the whole last name into an address. This address will then be automated using SMTP or python-O365.